requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0

# Interpretability and Visualization
shap>=0.42.0
lime>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
gradio>=3.40.0

# Image Processing and Analysis
albumentations>=1.3.0
imageio>=2.31.0
scikit-image>=0.21.0

# Database and Storage
sqlalchemy>=2.0.0
sqlite3
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
